Fri, Dec 22, 2023, 12:51 AM +00:00

+++ cs50_introduction/week_1/set/credit/credit.c
@@ -30,7 +30,7 @@ int main(void)
     {
         if (i % 2 == 1)
         {
-            int tmp;
+            long tmp;
             tmp = calc % 10;
             tmp *= 2;
             for (int j = 0; 0 < tmp; j++)
